+++ descqa/configs/quick_bkg_pk.yaml
@@ -1,0 +1,7 @@
+subclass_name: ImgPkTest.ImgPkTest
+description: 'Compute power spectrum of the background' 
+included_by_default: true
+
+input_path: '/global/cscratch1/sd/jsanch87/quick_pk_bkg/Pk_33.fits.gz'
+val_label: 'Test 33'
+raft: 'R14'
